refactor(engine): game text
game text, font, refactoring
